Implement comprehensive SEO metadata, robots.txt, and sitemap
