+++ src/app/robots.ts
@@ -1,0 +1,11 @@
+import { MetadataRoute } from "next";
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+    },
+    sitemap: "https://dpkreativ.vercel.app/sitemap.xml",
+  };
+}
